add check to prevent re-running test when data exists. work in progress

+++ scripts/dev/test_rt_kernel.sh
@@ -52,8 +52,12 @@ RT_PERIOD=100
 RATE=$(expr 1000 / $RT_PERIOD) # Convert RT period to freq in kHz
 
 # Run latency test under dynamic load
-stress --cpu 2 --vm 1 --hdd 1 --timeout $TIME & 
-sudo /usr/xenomai/bin/./latency -s -h -p $RT_PERIOD -B 1 -H 500000 -T $TIME -g test_rt_histdata.txt | tee test_rt_kernel.log
+if [ -f test_rt_histdata.txt ]; then
+	echo 'test run already'
+else
+	stress --cpu 2 --vm 1 --hdd 1 --timeout $TIME & 
+	sudo /usr/xenomai/bin/./latency -s -h -p $RT_PERIOD -B 1 -H 500000 -T $TIME -g test_rt_histdata.txt | tee test_rt_kernel.log
+fi
 
 sudo Rscript makeHistPlot.r "$DISTRO" "$HOSTNAME" "$RT_KERNEL" "$PROCESSOR" "$GRAPHICS_CARD" "$GRAPHICS_DRIVER" "$RATE" "$DAQ"
